Refactoring: moving plot3d toolbars to tools package
