+++ silx/gui/plot3d/Plot3DWindow.py
@@ -34,9 +34,9 @@ __date__ = "26/01/2017"
 
 from silx.gui import qt
 
-from .Plot3DToolBar import Plot3DOutputToolBar, Plot3DInteractiveModeToolBar
 from .Plot3DWidget import Plot3DWidget
-from .ViewpointToolBar import ViewpointToolBar
+from .tools import OutputToolBar, InteractiveModeToolBar
+from .tools import ViewpointToolBar
 
 
 class Plot3DWindow(qt.QMainWindow):
@@ -51,7 +51,7 @@ class Plot3DWindow(qt.QMainWindow):
         self._plot3D = Plot3DWidget()
         self.setCentralWidget(self._plot3D)
 
-        toolbar = Plot3DInteractiveModeToolBar(parent=self)
+        toolbar = InteractiveModeToolBar(parent=self)
         toolbar.setPlot3DWidget(self._plot3D)
         self.addToolBar(toolbar)
         self.addActions(toolbar.actions())
@@ -59,7 +59,7 @@ class Plot3DWindow(qt.QMainWindow):
         self.addToolBar(
             ViewpointToolBar(parent=self, plot3D=self._plot3D))
 
-        toolbar = Plot3DOutputToolBar(parent=self)
+        toolbar = OutputToolBar(parent=self)
         toolbar.setPlot3DWidget(self._plot3D)
         self.addToolBar(toolbar)
         self.addActions(toolbar.actions())
