+++ silx/gui/plot3d/tools/ViewpointToolBar.py
